+++ Core/Repository/TagsService.php
@@ -3,6 +3,8 @@
 namespace Netgen\TagsBundle\Core\Repository;
 
 use eZ\Publish\API\Repository\Repository;
+use eZ\Publish\API\Repository\Values\User\User;
+use eZ\Publish\API\Repository\Values\ValueObject;
 use Netgen\TagsBundle\API\Repository\TagsService as TagsServiceInterface;
 use Netgen\TagsBundle\SPI\Persistence\Tags\Handler as TagsHandler;
 use eZ\Publish\SPI\Persistence\Content\Language\Handler as LanguageHandler;
@@ -23,6 +25,7 @@ use eZ\Publish\Core\Base\Exceptions\InvalidArgumentException;
 use eZ\Publish\Core\Base\Exceptions\InvalidArgumentValue;
 use DateTime;
 use Exception;
+use Closure;
 
 class TagsService implements TagsServiceInterface
 {
@@ -40,6 +43,13 @@ class TagsService implements TagsServiceInterface
      * @var \eZ\Publish\SPI\Persistence\Content\Language\Handler
      */
     protected $languageHandler;
+
+    /**
+     * Counter for the current sudo nesting level.
+     *
+     * @var int
+     */
+    protected $sudoNestingLevel = 0;
 
     /**
      * Constructor.
@@ -72,7 +82,7 @@ class TagsService implements TagsServiceInterface
      */
     public function loadTag($tagId, array $languages = null, $useAlwaysAvailable = true)
     {
-        if ($this->repository->hasAccess('tags', 'read') !== true) {
+        if ($this->hasAccess('tags', 'read') !== true) {
             throw new UnauthorizedException('tags', 'read');
         }
 
@@ -99,7 +109,7 @@ class TagsService implements TagsServiceInterface
      */
     public function loadTagByRemoteId($remoteId, array $languages = null, $useAlwaysAvailable = true)
     {
-        if ($this->repository->hasAccess('tags', 'read') !== true) {
+        if ($this->hasAccess('tags', 'read') !== true) {
             throw new UnauthorizedException('tags', 'read');
         }
 
@@ -127,7 +137,7 @@ class TagsService implements TagsServiceInterface
      */
     public function loadTagByUrl($url, array $languages)
     {
-        if ($this->repository->hasAccess('tags', 'read') !== true) {
+        if ($this->hasAccess('tags', 'read') !== true) {
             throw new UnauthorizedException('tags', 'read');
         }
 
@@ -193,7 +203,7 @@ class TagsService implements TagsServiceInterface
      */
     public function loadTagChildren(Tag $tag = null, $offset = 0, $limit = -1, array $languages = null, $useAlwaysAvailable = true)
     {
-        if ($this->repository->hasAccess('tags', 'read') !== true) {
+        if ($this->hasAccess('tags', 'read') !== true) {
             throw new UnauthorizedException('tags', 'read');
         }
 
@@ -226,7 +236,7 @@ class TagsService implements TagsServiceInterface
      */
     public function getTagChildrenCount(Tag $tag = null, array $languages = null, $useAlwaysAvailable = true)
     {
-        if ($this->repository->hasAccess('tags', 'read') !== true) {
+        if ($this->hasAccess('tags', 'read') !== true) {
             throw new UnauthorizedException('tags', 'read');
         }
 
@@ -252,7 +262,7 @@ class TagsService implements TagsServiceInterface
      */
     public function loadTagsByKeyword($keyword, $language, $useAlwaysAvailable = true, $offset = 0, $limit = -1)
     {
-        if ($this->repository->hasAccess('tags', 'read') !== true) {
+        if ($this->hasAccess('tags', 'read') !== true) {
             throw new UnauthorizedException('tags', 'read');
         }
 
@@ -279,7 +289,7 @@ class TagsService implements TagsServiceInterface
      */
     public function getTagsByKeywordCount($keyword, $language, $useAlwaysAvailable = true)
     {
-        if ($this->repository->hasAccess('tags', 'read') !== true) {
+        if ($this->hasAccess('tags', 'read') !== true) {
             throw new UnauthorizedException('tags', 'read');
         }
 
@@ -302,7 +312,7 @@ class TagsService implements TagsServiceInterface
      */
     public function loadTagSynonyms(Tag $tag, $offset = 0, $limit = -1, array $languages = null, $useAlwaysAvailable = true)
     {
-        if ($this->repository->hasAccess('tags', 'read') !== true) {
+        if ($this->hasAccess('tags', 'read') !== true) {
             throw new UnauthorizedException('tags', 'read');
         }
 
@@ -340,7 +350,7 @@ class TagsService implements TagsServiceInterface
      */
     public function getTagSynonymCount(Tag $tag, array $languages = null, $useAlwaysAvailable = true)
     {
-        if ($this->repository->hasAccess('tags', 'read') !== true) {
+        if ($this->hasAccess('tags', 'read') !== true) {
             throw new UnauthorizedException('tags', 'read');
         }
 
@@ -369,7 +379,7 @@ class TagsService implements TagsServiceInterface
      */
     public function getRelatedContent(Tag $tag, $offset = 0, $limit = -1)
     {
-        if ($this->repository->hasAccess('tags', 'read') !== true) {
+        if ($this->hasAccess('tags', 'read') !== true) {
             throw new UnauthorizedException('tags', 'read');
         }
 
@@ -409,7 +419,7 @@ class TagsService implements TagsServiceInterface
      */
     public function getRelatedContentCount(Tag $tag)
     {
-        if ($this->repository->hasAccess('tags', 'read') !== true) {
+        if ($this->hasAccess('tags', 'read') !== true) {
             throw new UnauthorizedException('tags', 'read');
         }
 
@@ -446,10 +456,10 @@ class TagsService implements TagsServiceInterface
         $keywords = $tagCreateStruct->getKeywords();
 
         if (!empty($tagCreateStruct->parentTagId)) {
-            if ($this->repository->canUser('tags', 'add', $this->loadTag($tagCreateStruct->parentTagId)) !== true) {
+            if ($this->canUser('tags', 'add', $this->loadTag($tagCreateStruct->parentTagId)) !== true) {
                 throw new UnauthorizedException('tags', 'add');
             }
-        } elseif ($this->repository->hasAccess('tags', 'add') !== true) {
+        } elseif ($this->hasAccess('tags', 'add') !== true) {
             throw new UnauthorizedException('tags', 'add');
         }
 
@@ -521,11 +531,11 @@ class TagsService implements TagsServiceInterface
         $keywords = $tagUpdateStruct->getKeywords();
 
         if ($tag->mainTagId > 0) {
-            if ($this->repository->hasAccess('tags', 'edit') !== true) {
+            if ($this->hasAccess('tags', 'edit') !== true) {
                 throw new UnauthorizedException('tags', 'edit');
             }
         } else {
-            if ($this->repository->hasAccess('tags', 'editsynonym') !== true) {
+            if ($this->hasAccess('tags', 'editsynonym') !== true) {
                 throw new UnauthorizedException('tags', 'editsynonym');
             }
         }
@@ -613,7 +623,7 @@ class TagsService implements TagsServiceInterface
     {
         $keywords = $synonymCreateStruct->getKeywords();
 
-        if ($this->repository->hasAccess('tags', 'addsynonym') !== true) {
+        if ($this->hasAccess('tags', 'addsynonym') !== true) {
             throw new UnauthorizedException('tags', 'addsynonym');
         }
 
@@ -688,7 +698,7 @@ class TagsService implements TagsServiceInterface
      */
     public function convertToSynonym(Tag $tag, Tag $mainTag)
     {
-        if ($this->repository->hasAccess('tags', 'makesynonym') !== true) {
+        if ($this->hasAccess('tags', 'makesynonym') !== true) {
             throw new UnauthorizedException('tags', 'makesynonym');
         }
 
@@ -736,7 +746,7 @@ class TagsService implements TagsServiceInterface
      */
     public function mergeTags(Tag $tag, Tag $targetTag)
     {
-        if ($this->repository->hasAccess('tags', 'merge') !== true) {
+        if ($this->hasAccess('tags', 'merge') !== true) {
             throw new UnauthorizedException('tags', 'merge');
         }
 
@@ -785,7 +795,7 @@ class TagsService implements TagsServiceInterface
      */
     public function copySubtree(Tag $tag, Tag $targetParentTag)
     {
-        if ($this->repository->hasAccess('tags', 'read') !== true) {
+        if ($this->hasAccess('tags', 'read') !== true) {
             throw new UnauthorizedException('tags', 'read');
         }
 
@@ -836,7 +846,7 @@ class TagsService implements TagsServiceInterface
      */
     public function moveSubtree(Tag $tag, Tag $targetParentTag)
     {
-        if ($this->repository->hasAccess('tags', 'edit') !== true) {
+        if ($this->hasAccess('tags', 'edit') !== true) {
             throw new UnauthorizedException('tags', 'edit');
         }
 
@@ -884,11 +894,11 @@ class TagsService implements TagsServiceInterface
     public function deleteTag(Tag $tag)
     {
         if ($tag->mainTagId > 0) {
-            if ($this->repository->hasAccess('tags', 'deletesynonym') !== true) {
+            if ($this->hasAccess('tags', 'deletesynonym') !== true) {
                 throw new UnauthorizedException('tags', 'deletesynonym');
             }
         } else {
-            if ($this->repository->hasAccess('tags', 'delete') !== true) {
+            if ($this->hasAccess('tags', 'delete') !== true) {
                 throw new UnauthorizedException('tags', 'delete');
             }
         }
@@ -972,5 +982,81 @@ class TagsService implements TagsServiceInterface
                 'languageCodes' => $languageCodes,
             )
         );
+    }
+
+    /**
+     * Allows tags API execution to be performed with full access sand-boxed.
+     *
+     * The closure sandbox will do a catch all on exceptions and rethrow after
+     * re-setting the sudo flag.
+     *
+     * Example use:
+     *     $tag = $tagsService->sudo(
+     *         function (TagsService $tagsService) use ($tagId) {
+     *             return $tagsService->loadTag($tagId);
+     *         }
+     *     );
+     *
+     * @param \Closure $callback
+     * @param \Netgen\TagsBundle\API\Repository\TagsService $outerTagsService
+     *
+     * @throws \RuntimeException Thrown on recursive sudo() use.
+     * @throws \Exception Re throws exceptions thrown inside $callback
+     *
+     * @return mixed
+     */
+    public function sudo(Closure $callback, TagsServiceInterface $outerTagsService = null)
+    {
+        ++$this->sudoNestingLevel;
+        try {
+            $returnValue = $callback($outerTagsService !== null ? $outerTagsService : $this);
+        } catch (Exception $e) {
+            --$this->sudoNestingLevel;
+            throw $e;
+        }
+
+        --$this->sudoNestingLevel;
+
+        return $returnValue;
+    }
+
+    /**
+     * Checks if user has access to specified module and function.
+     *
+     * @param string $module The module, aka controller identifier to check permissions on
+     * @param string $function The function, aka the controller action to check permissions on
+     * @param \eZ\Publish\API\Repository\Values\User\User $user
+     *
+     * @return bool|array if limitations are on this function an array of limitations is returned
+     */
+    public function hasAccess($module, $function, User $user = null)
+    {
+        // Full access if sudo nesting level is set by sudo method
+        if ($this->sudoNestingLevel > 0) {
+            return true;
+        }
+
+        return $this->repository->hasAccess($module, $function, $user);
+    }
+
+    /**
+     * Indicates if the current user is allowed to perform an action given by the function on the given
+     * objects.
+     *
+     * @param string $module The module, aka controller identifier to check permissions on
+     * @param string $function The function, aka the controller action to check permissions on
+     * @param \eZ\Publish\API\Repository\Values\ValueObject $object The object to check if the user has access to
+     * @param mixed $targets The location, parent or "assignment" value object, or an array of the same
+     *
+     * @return bool
+     */
+    public function canUser($module, $function, ValueObject $object, $targets = null)
+    {
+        $permissionSets = $this->hasAccess($module, $function);
+        if ($permissionSets === false || $permissionSets === true) {
+            return $permissionSets;
+        }
+
+        return $this->repository->canUser($module, $function, $object, $targets);
     }
 }