+++ Resources/doc/CHANGELOG.md
@@ -1,6 +1,12 @@
 Netgen Tags Bundle changelog
 ============================
 
+2.0.7 (28.10.2015)
+------------------
+
+* Add `sudo` method to Tags service, used by tags router to match the tag since at that point user is still anonymous
+* Add setting to enable or disable tag router, useful for legacy siteaccesses
+
 2.0.6 (08.10.2015)
 ------------------